Working read

+++ src/main/java/com/qa/choonz/persistence/repository/AlbumRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.qa.choonz.persistence.domain.Album;
 
 @Repository
-public interface AlbumRepository extends JpaRepository<Album, Long> {
+public interface AlbumRepository extends JpaRepository<Album, Integer> {
 
 }
